Services, Serviceslmpl

+++ Back1/gadi/src/main/java/com/ista/gadi/Dao/I_Actividad_Dao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ista.gadi.Entity.Actividad;
 
-public interface I_Actividad_Dao extends CrudRepository<Actividad, Integer>{
+public interface I_Actividad_Dao extends CrudRepository<Actividad, Long>{
 	
 }
