+++ Back1/gadi/src/main/java/com/ista/gadi/Dao/I_Asignatura_Dao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ista.gadi.Entity.Asignatura;
 
-public interface I_Asignatura_Dao extends CrudRepository<Asignatura, Integer>{
+public interface I_Asignatura_Dao extends CrudRepository<Asignatura, Long>{
 
 }
