+++ Back1/gadi/src/main/java/com/ista/gadi/Dao/I_Carrera_Dao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ista.gadi.Entity.Carrera;
 
-public interface I_Carrera_Dao extends CrudRepository<Carrera, Integer>{
+public interface I_Carrera_Dao extends CrudRepository<Carrera, Long>{
 
 }
