+++ Back1/gadi/src/main/java/com/ista/gadi/Dao/I_Ciclo_Dao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ista.gadi.Entity.Ciclo;
 
-public interface I_Ciclo_Dao extends CrudRepository<Ciclo, Integer>{
+public interface I_Ciclo_Dao extends CrudRepository<Ciclo, Long>{
 
 }
