+++ Back1/gadi/src/main/java/com/ista/gadi/Dao/I_Distributivo_Dao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ista.gadi.Entity.Distributivo;
 
-public interface I_Distributivo_Dao extends CrudRepository<Distributivo, Integer>{
+public interface I_Distributivo_Dao extends CrudRepository<Distributivo, Long>{
 
 }
