+++ Back1/gadi/src/main/java/com/ista/gadi/Dao/I_Jornada_Dao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ista.gadi.Entity.Jornada;
 
-public interface I_Jornada_Dao extends CrudRepository<Jornada, Integer>{
+public interface I_Jornada_Dao extends CrudRepository<Jornada, Long>{
 
 }
