+++ Back1/gadi/src/main/java/com/ista/gadi/Dao/I_Periodo_Dao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ista.gadi.Entity.Periodo;
 
-public interface I_Periodo_Dao extends CrudRepository<Periodo, Integer>{
+public interface I_Periodo_Dao extends CrudRepository<Periodo, Long>{
 
 }
