+++ Back1/gadi/src/main/java/com/ista/gadi/Dao/I_Tipo_Actividad_Dao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ista.gadi.Entity.Tipo_Actividad;
 
-public interface I_Tipo_Actividad_Dao extends CrudRepository<Tipo_Actividad, Integer>{
+public interface I_Tipo_Actividad_Dao extends CrudRepository<Tipo_Actividad, Long>{
 
 }
